task: organization of endpoints
task: organization of endpoints
