+++ src/endpoints/mod.rs
@@ -1,6 +1,8 @@
 pub mod endpoint_settings;
 pub mod endpoint_cli;
-mod commands;
+mod run_cli;
+mod saved_cli;
+mod utils;
 
 #[cfg(test)]
 mod tests;
